requirements.txt
streamlit==1.28.2
streamlit-webrtc==0.45.0
opencv-python-headless==4.9.0.80
numpy==1.23.3
pillow
av==10.0.0

